refactor(codegen): extrai calls/new_expr.rs
Move 4 fns relacionadas a `new` (~709 LOC) para
expressions/calls/new_expr.rs:
- lower_new (pub(crate) re-export): aloca instancia GC + chama
  __class_C__init + constructor
- lower_new_function: `new Function("body")` via runtime.eval
- lower_function_handle_method: chamadas em handles Function
  reificados (.call/.apply/.bind/etc)
- lower_function_method_call: dispatch de metodos sobre handle Function

calls/mod.rs: 2521 -> 1818 linhas. Suite verde.
